requirements.txt
PyQt6>=6.4.0
psutil>=5.9.0
requests>=2.28.0
python-dotenv>=0.19.0
SQLAlchemy>=1.4.0
PyYAML>=6.0
rich>=12.0.0
